requirements.txt
Django==1.8
Pillow==2.8.1
dj-database-url==0.3.0
django-dotenv==1.3.0
django-modeltranslation==0.9
django-mptt==0.7.2
django-rosetta==0.7.6
microsofttranslator==0.5
polib==1.0.6
psycopg2==2.6
requests==2.6.2
six==1.9.0
wsgiref==0.1.2
